+++ addon/components/render-content.js
@@ -1,18 +1,12 @@
 import Component from '@ember/component';
-import { computed } from '@ember/object';
 import layout from '../templates/components/render-content';
 
 export default Component.extend({
-  tagName: '',
-
   layout,
 
-  content: null,
+  tagName: '',
 
-  contentIsComponent: computed('content', function() {
-    let contentConstructorName = this.get('content.constructor.name') || '';
-    return contentConstructorName.indexOf('ComponentDefinition') > -1;
-  }).readOnly(),
+  content: null,
 }).reopenClass({
   positionalParams: ['content'],
 });